single precision

+++ src/fitting.jl
@@ -99,7 +99,7 @@ const default_param_bounds = Dict(:A => (0.1, 50000), :λ => (0.01, 0.5), :α =>
 # Keyword Arguments
 Requirements differ by model type...
 * `ButlerVolmer`, `AsymptoticMarcusHushChidsey`, `Marcus`: none
-* `MarcusHushChidsey`: average_dos::Float64 OR dos::DOSData OR dos_file::String
+* `MarcusHushChidsey`: average_dos::Float32 OR dos::DOSData OR dos_file::String
 * `MarcusHushChidseyDOS`: dos::DOSData OR dos_file
 Some are always options...
 * `param_bounds::Dict{Symbol,Any}`: ranges of guesses for relevant model parameters. (must include all necessary keys, but defaults to some sensible ranges if not provided, see `default_param_bounds`...note that you should provide this for faster fitting if you know bounds)
@@ -164,8 +164,8 @@ function _fit_model(
     #         s[i] = gs[i]
     #     end
     # end
-    lower = Float64.([param_bounds[p][1] for p in fitting_params(model_type)])
-    upper = Float64.([param_bounds[p][2] for p in fitting_params(model_type)])
+    lower = Float32.([param_bounds[p][1] for p in fitting_params(model_type)])
+    upper = Float32.([param_bounds[p][2] for p in fitting_params(model_type)])
     init_guess = 0.5 .* (lower .+ upper)
 
     # set optimisers based on the model type
